requirements.txt
scikit-learn
pandas
numpy
requests
scapy
streamlit
plotly
sqlite-utils
python-dotenv
APScheduler